Two import bug fixes (fixed by Olivier B)
* ValueError: ('invalid number of arguments', mul(Subtensor{0}.0, 18))
This error happened after serializing a model. Scalar ops taking a variable
number of arguments had their ufunc set to that op with -1 input parameters, and
nout output params. Code has been added to check this condition and set ufunc to
None which postpones the building of the ufunc to when we know the actual number of
params (or something like that -see OlivierB.).

* Optimization bug which seemed rather serious.
Fixed bug with Canonizer and DimShuffle
Symptom: ValueError: ('Input dimensions do not match.'