ConvOpGrad in full mode is currently not supported. Raise a NotImplementedError.
This is because the gradient on the weights is obtained by convoling the visible
layer with the hidden layer (as the kernel). Problem is the convolution code
does not support kernels larger than the input image (which is the case in full
mode). Either fix ConvOp to allow this, or maybe convolve the hidden with the
visible as the kernel, then DimShuffle at the end.