Switched the default strictness value of shared() back to strict=False.
Reason being there is a problem with the way we deal with strict.  Strictness
requires that a value has exactly the right type, but non-strictness implies
just about any kind of casting including in-exact casting.  What I would like
the default to be is something that means casting is ok as long as it is exact.