Replaced tensor.basic.SetSubtensor with tensor.basic.IncSubTensor.
This was done in order to greatly simplify a new optimization to serialize
multiple IncSubtensors of a common base.  This is a case that happens if you
compute more than one subtensor from a single variable and then take a gradient.
Serialization helps because then all the IncSubtensors can be done inplace.