Added tensor.autocast_int and tensor.autocast_float
These two functions are used by as_tensor_variable to determine how to turn
python ints and floats into ndarrays for TensorConstants.

This provides an i-hope-not-too-hacky way for config.floatX=='float32' to make it so
that python literals like 1.1 don't force an upcast in expressions like
(fvector() + 1.1).

Another option would have been to leave the downcast of 1.1 in the graph as a
symbolic node that would be pre-computed at compile time, but I think that would
behave pretty similarly, and further burden the optimizer.