added scalar and elemwise tensor op maximum and mininum and test for them.
numpy define the fct maximum and mininum as doing elemwise max and min.
We have the theano.tensor.{largest, smallest} that did that but with more op.
That make the graph harder to read. We will probably map largest and smallest
to the new fct in the futur after time show they are good.