nvcc - added post-processing of stderr from nvcc to remove warnings about things we do on purpose.
Also moved printing of the remainder of the message to INFO level logging, which
might be a bad idea.
But nvcc spews things that are not errors to stderr which are not easy to filter
out.  I'm thinking of the warnings that _POSIX_SOURCE is redefined, followed by
a multiline header traceback.
We do not want to print that to stderr by default every time someone compiles
anything with nvcc.

If compilation fails, then the whole nvcc stderr goes to our stderr anyway.