fixed tests of shared_randomstreams and randomstreams
The test_default_shape function in both modules was coded incorrectly. It used
the same generator for uniform and multinomial samples, whereas the modules being
tested used different generators for the uniform an multinomial samples.

It only worked for as long as it did because only a single multinomial sample
was being tested, and by chance the different generators  happened to draw the
same bit.