Refactor fusion and make it fuse one more case.
Their was duplicat code to test the maximum number of parameter to a GpuElemwise op.
When fusing GpuElemwise with a new inputs, we check that we don't bust the maximum number of inputs for that ops. In the past, we make the check after and their where case where we could have fused in some of the inputs.