added scalar.identity to local_useless_elemwise
I am not sure why this is necessary now.  One guess is that the canonicalizer or
something else has been modified to put in identity(), and now it is more
obvious that we need an optimization to take it out again.  I'm hoping that this
improves the slowdown recently observed in convolutional net benchmark.