test gpu elemwise fusion only on 3d tensor instead of 4d.
This fix the remaining case error in the buildbot. This don't hide the new case fused in the previous commit as their was error before it in the gpu 3d case.

The error in 4d are not real error as they are caused by limit in the number of input in a gpu elemwise op we can put in the gpu code that exist now.