Refactor tests in order to check consistency at each step, avoiding divergence.
Tests used to fail due to accumulation of negligible errors, so the
differences were big after a few iterations. Instead, we compare the results
at the end of each step, and synchronize the parameter values so that both
implementation restart from the same point.
Also, tightened the relative error bound.