Fix KnownFailure relative to gemm (and partially solves #415).
This does not change the logic of the Gemm optimizer, so #415 might
still be needed in the future.

This change:
  - apply the following optimization whether or not the transpose is inplace:
T.dot(X,Y).T -> T.dot(Y.T, X.T). The latter form is expected by the Gemm
optimizer.
  - removes the Warning class raised in test_blas.py:just_gemm, because
it was caught silently, hiding the fact that another case was failing.
A "Failure" is now thrown instead;
  - adds another test case for transposition;
  - makes the first case in test_blas.py:test_inplace0 expect a
"gemm_no_inplace" instead of _dot22.