Cache optimization: modules are not duplicated anymore in the cache
The idea is that even when the key changes, the resulting code and compiled
module can remain the same. We avoid this by keeping a hash of the C code (as
well as of the compilation options) and when this hash matches that of an
existing module, we re-use this module instead of having two different modules
doing the same thing.

Two improvements should be made to the current implementation:
    1. Currently a module is still compiled before we detect that it is
       duplicated, while we should avoid this compilation step.
    2. It may be the case that different code or compilation options actually
       lead to the exact same compiled module. It would be nice to be able to
       detect this situation as well.