Fix to the provided memory bug with truncated BPTT.
If scan receives memory in the output storage map it uses it. Now if you run
truncated BPTT, you do not loop over the entire sequence, and you should
return 0 for all the entries you don not visit. I used to assume that the
data I allocate does that ( because I used to allocate it using zeros), but
now if data is provided I did not check if those values are actually 0s.