Modified the way the version of lazylinker is checked to avoid problems under Windows.
The problem was that the version was obtained *after* importing the C module. As a result,
recompilation would fail because the library was still in use.
This change also saves the version in the __init__.py file of the lazylinker_ext base module,
so that it can be checked without loading the C module.