redefined tensor.basic.outer in terms of dot and dimshuffle.
Rationale: this way it can be optimized with the same machinery that currently
targets dot.

I will add back a modified version of Outer as sandbox/blas_ger.py. In future,
we should have a GER Op in blas.py, with a scipy and/or c implementation that
uses BLAS's sger/dger/cger/zger functions.