Increase the max_ratio of equilibrium
My optimization pushes in the main graph a lot of computations in terms of
constants, and constant folding maxes out pretty quickly. This is a quick
fix to that. I'm not sure if 50 is not too much, I haven't played with that
value .. maybe 20 or something would suffice. I would also like to increase
this value only for constant folding and not all the others, but I didn't
know how.