transformed the pushout optimization from local to a global opt.
There is no reason for this, just that I had a lot of problems with this
optimization. It turned out the problem to be caused by
local_useless_subtensor optimization (fix which I will push later).
In my attempts to debug it though, I ended up converting it into a global
optimizer, and I don't want to go through the pain of reverting it and
making sure also that it works.