Fixed a bunch of errors in equal_computation
Errors are obvious, first of all the data of the constants don't have to
scalars, and secondly, is not enough for outputs that have no owner to have
same type, but they should have the same index (i.e. correspond to the same
input). As I have all corresponding pairs of inputs in `common` and these
outputs are equal with their inputs, is sufficient to check if them ( as
pair) are part of common