Fixed bug in scan_perform.pyx
Scan_perform.pyx used not to zero out the output starting from the right
point (when performing a while). This bug was detectable only in debug mode.
The python version of scan was doing the right thing though, so this is just
a matter of sync-ing the two versions. Note that all the changes in
scan_perform.c where generated by cython, and as such they shouldn't need to
be checked ( they are mostly because in the lab we have cython 0.14 instead
of cython 0.13 which I used previously).