make sure that in speed tests for RNN everything runs with CVM.
I wanted to make sure that I get the right timings. It seems that CVM is
faster then what James suggested for now on CPU, and slower on GPU(but can
be fixed). Note that a full implementation of what James suggested will be
slower then what this text predicts ( this is as fast as it will ever get ),
namely scan needs to do a bit more bookeeping that the cython right now
does.