Deprecated tensor.shared (renamed tensor._shared)
Most of the time people who use tensor.shared do it because they mix it up
with theano.shared. By renaming it with a leading underscore, this confusion
should disappear.

By the way this commit also changes a few calls in Theano code from
tensor.shared to theano.shared, when I thought calling tensor.shared was not
intentional.