Removed keep_wrt_type parameter of tensor.grad
The reason I had added this parameter in 0b8b021 was to let the user get rid of
annoying warnings. However, I probably shouldn't have given the option to keep
the old behavior with no warning, since this behavior was unintuitive and
potentially dangerous. Since the warning was removed by 965afb6, it seems
pointless to keep this extra parameter that adds extra complexity to the
tensor.grad interface with no benefit.