replacing transposed dot patternsub with local opt
The old form (PatternSub) had the problem that it wasn't calling broadcast_like
on the return value.  It would have been wrong to hack a broadcast_like into
gof.opt, so I rewrote these two optimizations as a function.

PatternSub is a really cool idea, in the future it might be worth writing a
TensorPatternSub that deals with tensor-specific issues, like casting,
dimshuffles, and rebroadcasts.