updated functions to extract different arguments of scan
Scan arguments are ordered, and depending on the index, an argument
can be a sequence, a state/output or a non sequence. The scan op has methods
that given the full list of arguments returns only those that represent a
certain category of inputs. The old functions used to look at self.inputs,
self.outputs or node.inputs/node.outputs to determine those entries. This
approach was restrictive, since all functions on scan most of the time work
on clones of the original arguments. The new functions now take a list of
arguments (could be either the original or clones).