added conventions to make_node
The make_node function used to check if inputs of the scan op have the same
type. In order to do so it has to parse them, This comments indicate the
convention in naming variables when this check is done (so that there is no
confusion between the variables representing the inputs of the lambda
functions, i.e. slices of the inputs and the actual inputs).