removed useless class
Before this class was used to validate that computing the shape of an
argument does not rely on any input of the inner function of scan. We do not
need to do this anymore, since now the output storage size is given as input
to scan, and output shapes are the same as input shape (so the internal
graph of scan does not come into play anymore).