new scan_permutation op
This op takes care of the special case when the entries of scan are not in
the righr order. Such a thing can happen once the save mem optimization is
implemented, and the memory buffer for an output is much smaller then the
number of steps. Then scan will just roll around that memory buffer, and
entry 0 might end up somewhere in the middle