reverted fixed pushed yesturday
The problem was that CudaNdarrays would not allow you to zero some part of
them. My fix was to allocate a ndarray and have then the CudaNdarray setitem
to copy that memory on the card. Fred fixed the bug on the CudaNdarray side
though so this fix (which is suboptimal) is not needed anymore.