Fixed bug in deque implementation for Python 2.4
The 'remove' method was bugged (it would not remove an item if it was the last
item in the queue).
This commit actually replaces the full custom deque implementation by a subclass
of the original Python 2.4 implementation of deque, as it seems safer than
re-writing everything.

The implementation is also moved to the python25.py file since it is the place
for Python 2.4 compatibility code.