Fixed gh-356: dtype of tensor.sum()
The default dtype for signed integer input types is now int64 (and
uint64 for unsigned integers).
The default dtype can be overridden by specifying a dtype manually.
However this does not appear to be very safe, so a warning was added to
the docstrings when doing so.

This commit also contains a few minor PEP8 fixes.