Fixed crash with Composite support code and GPU
The problem was that when a Composite Op had a Composite subnode, having
the subnode return an empty support code would cause the parent node to
return '\n' as support code. As a result, GPU code generation would
crash due to non-empty support code.

Thanks to Fred for writing the associated test.