fix for computing Rop wrt to shared variables in a scan
Issue : scan used to compute something, that I thought was the R-op with
respect to shared variables with updates of the inner function. After
more reflection, I think those numbers I'm computing have no meaning ..

Unless the user asks explicitly for the R_op with respect to some value
that is a shared variable, this would not affect anything (just some extra
computations). The reason I removed them is not only the possible
meaningless of these numbers, but also the fact that R_op methods are not
defined for random number sampling function, which would appear as
shared variable. So you would get that you can not compute the R-op because
some function is not defined for that op, even though that is only to
compute some values you don't care and which might not have any meaning.

I have no idea exactly how to test this, since is not a bug per-se is more
of a semanting error in my understanding of what the R-op is suppose to do.