More sane inplace optimization
The optimization before was all or nothing, which is a bad compromise for
scan. This new optimization tries to work in place on each outputs, and
keeps only those for which it can. This way you do not get your entire op to
be non-inplace because of a single output that can not be computed inplace.