Change merge feature to avoid looping through env.
Replacement for equivalent nodes and constants are now queued when
the graph is modified, through the MergeFeature, and processed when the
optimization is applied.
This means that appling the MergeOptimization when there is nothing
to be merged is now almost instantaneous, even for big graphs.
It is also faster when lots of nodes have to be merged (empirical 40 %
speed-up for a graph with ~ 5000 nodes).