fix context problem with pycuda by checking it is imported before.
Also add the pycuda.init flag to force its import.

It is not needed to set it to True in the buildbot, as all test file
get imported before test start to run. So this init PyCUDA first if
the theano flag device and init_gpu_device are not used.