Corected use of UPLO parameter in Eigh.grad.
Numpy's eigh(a, 'L') (eigh(a, 'U')) is a function of tril(a) (triu(a))
only.  This means that partial derivative of eigh(a, 'L') (eigh(a,
'U')) with respect to a[i,j] is zero for i < j (i > j).  At the same
time, non-zero components of the gradient must account for the fact
that variation of the opposite triangle contributes to variation of
two elements of Hermitian (symmetric) matrix.