Make check_preallocated_output work on fct outputs
In DebugMode, do not use the no_recycling list when calling the CLinker
to build individual thunks, so that output memory can always be re-used,
even for outputs of the actual function.

This enables check_preallocated_output to work on Ops that compute
the function's outputs.

This should close gh-1174.