Remove .fgraph attribute from Apply and Variable before pickling
A FunctionGraph cannot be pickled usually, because it has many
partial functions and lambda function as members.
Since the Composite Op needs to pickle variables and apply nodes that
belong to an fgraph, trying to cache C code of a composite op raised
an Exception, which means C code for composite ops had to be recompiled
at each execution.

The easiest way seemed to never pickle the .fgraph attribute of an Apply
or Variable Node.