Speed up MaxAndArgmaxOptimizer.
This is done by making it a local optimizer. This change to allow
local optimizer to remove variable in the graph. I allow this only for
variable that aren't used.

In a small test case, it wsa taking 0.6s now it take .1s. In a bigger
test case, it was taking 40s. I didn't rerun it with the opt as it is
too long.

This speed up the opt, as we do less iteration on the graph, as in the
EquilibriumOptimizer add in the current iteration the new node in the
graph. The old opt wasn't doing this and I didn't wanted to duplicate
all those type of code.