Remove useless debugmode warning about optimization that change results
The error is due to rounding difference. The test that fail randomly is theano/sandbox/cuda/tests/test_conv_cuda_ndarray.py:TestConv2DGPU.test_logical_shapes.

Normally, we should set atol depending of the input shape, but the current structure don't allow this, so I use the output shape. So this could happen with other tests if the output shape is smaller.