bugfix, for a case that don't happen right now.
The preallocated output size for make_vector should ALWAYS be the good
shape right now as it only depend on the ndim and that is constant in
the graph and we only give preallocated output of the same node.

This fix will be useful in the futur if we make other type of
preallocated output that could be of the wrong size.