(This was started with Raul Chandias Ferrari) Flatten Composite.
This help gh-689 as it allow it to support up to 57 iteration of the
loop instead of just 43.

This could also help Theano cache in the case where graph could have
different fusion order, so different way the composite are in each
other, but do the same computation.

This make the printing of Composite more readable.