Remove a file under Apache licence (argparse)
Argparse is in Python 2.7. This file was added for compatibility with
older python.

We remove one of its user(nose_pr.py) that we don't use.

As we support python 2.6, I raise a good error message in the only
file that use it. As it is available in pypi, I think it is fine we
don't provide it for python 2.6, as this file isn't very important.