[disabled BUGfix,CRASH] related to the cudnn pooling gradient.
bug: The optimization where converting the ignore_border=False to ignore_border=True.

crash: The optimization that convert the gpudownsample grad to
cuddpoolgrad was swapping the inputs parameter. This is caused in part
by cudnn using strange definition of input and output. Now the op use
the same inputs order as the other pool grad ops. The only difference
is tha the c_code() use the name that cudnn use. There is a comment
that tell that.