Maybe fix scan test?
I'm not sure what this test is expecting.  It seems to test that the gpu will magically work and optimize the graph even though we disable InputsToGpuOptimizer (which is obviously not the case).  I wonder why it passed before.  I wonder if this is the right fix.  I also wonder what is it testing exactly.

In any case it seems to pass now.