Revert all of that work, fix the comprehension error that lead to it
Instead of putting the "if you work harder you can get better performance"
note before the code, make the code the first thing you see and only
note that you could do it another way in the performance testing section.

Also switches to code-blocks throughout.

Adds note that the openblas thread limit note seems to be outdated,
the section above notes that we should be linked to multithreaded
blas already, and testing seems to suggest we are.