Add new scalar and elemwise op erfcx
op for calculating the scaled complementary error function
exp(x**2)*erfc(x) in a numerically stable way. This op is useful for
calculating various things related to Gaussian distributions, such as
truncated moments and expressions involving the log(cdf(x)) for large
negative values of x. The Theano code closely follows that of the Erfinv
op.