Small visual changes (blue color)
I’ve added the key ‘Output’ to the dict colorCodes in order to be able
to changes the ‘blue’ color of Theano graph node which is not well
visible. Personally, I used the 'Output': '#87CEFA'  color code calling
pydotprint(…).

An alternative correction should be to replace ‘blue’ by '#87CEFA’.
It’s up to you…