Fix abstractconv_grad dilation and enable border_mode tests.
This fixes a typo in bbc749ab516, which would give problems with
non-symmetric filter dilations for border_modes full or half. (This
error was only visible in the debug mode.)

This commit also fixes the test, which previously would not actually
test different border_modes, but would always use the same default.

Finally, this commit adds border_mode = "half" to the tests.