Delete the tests related to fftconv optimisations.
They've been broken for a while now and we only noticed now since we
installed scikits-cuda on the test machine.  If people really want to
use this, they can dig though the code and call it directly.  The
functional tests are still there and the pass so that's good enough
for me.