Update. Many of @abergeron comments have been taken account.
Tests have been rewritten, and they run so that
a CPU computation and a GPU computation are always
performed (separately) on the same input.

This allow to run the tests with Theano profiling
flags and then compare the execution time of
MaxAndArgmax (CPU) and GpuMaxAndArgmax (GPU).

Some code has also been modified in theano/tensor/basic.py,
related to MaxAndArgmax, to make the API more uniform
and to put most of axis checking in maxandargmax wrapper
instead of in make_node functions of (Gpu)MaxAndArgmax.