Fix Windows issue related to PR #5207 ( see comment https://github.com/Theano/Theano/pull/5207#issuecomment-260430352 ).
In Windows there are some issues with
os.listdir(path) when path is surrounded
with quotes. So it's better to remove
quotes (if present) at the start and the end
of paths passed to this function.

Checking of ldflags parts is also modified
to better handle quotes around "-L" args.