Enable L2 tests
Nose test collects tests recursively, so it is sensitive to the $PWD
at the time you run it.

Also it seems there's no need for both the verbose test AND the
batch test, so I commented out the batch version since it appears it
would just makes the log less useful without taking less time.